fix: 애플리케이션, API 키 간의 트랜잭션 오류 수정
1. 앱 생성 -> apikey 삭제 -> apikey 생성시 오류
- 앱과 키 사이의 1:1 관계 때문에 트랜잭션 오류 발생
- 1:N 관계로 수정해서 해결 (정책도 일단 1:N으로)

2. 1번 상태 (app 존재 apikey 없는 상태) -> app 삭제시 오류
앱 삭제시 키가 삭제되어있는지에 대한 조건문이 없었기에 에러 발생

+++ app/services/application_service.py
@@ -155,10 +155,11 @@ class ApplicationService:
                 detail="애플리케이션을 찾을 수 없습니다."
             )
 
+        # 4. 연결된 API 키를 삭제합니다.
+        if key:
+            self.apiKeyRepo.delete_key(key.id)
+
         # 3. 애플리케이션을 소프트 삭제합니다.
         self.appRepo.delete_application(appId)
-
-        # 4. 연결된 API 키를 삭제합니다.
-        self.apiKeyRepo.delete_key(key.id)
 
         return self.map_to_application_response(app, key)
